feat(cli): add feature to disable downloads
