+++ packages/cli/src/tools/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod wasm_bindgen;
